Fix numpy type not in openapi schema

+++ src/main/python/dynamic_hosting/example_model_training/miniloan-rfc.py
@@ -101,12 +101,12 @@ def main():
             {
                 'name': "creditScore",
                 'order': 0,
-                'type': 'int'
+                'type': 'int64'
             },
             {
                 'name': "income",
                 'order': 1,
-                'type': 'float64'
+                'type': 'long'
             },
             {
                 'name': "loanAmount",
